fix: click on field in the list item (#3120)
