--- frontend/src/container/LogExplorerDetailedView/LogExplorerDetailedView.interfaces.ts
@@ -1,6 +1,0 @@
-import { ILog } from 'types/api/logs/log';
-
-export type LogExplorerDetailedViewProps = {
-	log: ILog | null;
-	onClose: () => void;
-};
