+++ frontend/src/container/LogsExplorerList/LogsExplorerList.interfaces.ts
@@ -1,3 +1,4 @@
+import { AddToQueryHOCProps } from 'components/Logs/AddToQueryHOC';
 import { ILog } from 'types/api/logs/log';
 import { IBuilderQuery } from 'types/api/queryBuilder/queryBuilderData';
 
@@ -8,4 +9,4 @@ export type LogsExplorerListProps = {
 	onEndReached: (index: number) => void;
 	onExpand: (log: ILog) => void;
 	onOpenDetailedView: (log: ILog) => void;
-};
+} & Pick<AddToQueryHOCProps, 'onAddToQuery'>;
