+++ frontend/src/container/LogsExplorerList/index.tsx
@@ -32,6 +32,7 @@ function LogsExplorerList({
 	onOpenDetailedView,
 	onEndReached,
 	onExpand,
+	onAddToQuery,
 }: LogsExplorerListProps): JSX.Element {
 	const { initialDataSource } = useQueryBuilder();
 
@@ -79,6 +80,7 @@ function LogsExplorerList({
 					logData={log}
 					selectedFields={selectedFields}
 					onOpenDetailedView={onOpenDetailedView}
+					onAddToQuery={onAddToQuery}
 				/>
 			);
 		},
@@ -87,6 +89,7 @@ function LogsExplorerList({
 			options.maxLines,
 			selectedFields,
 			onOpenDetailedView,
+			onAddToQuery,
 			onExpand,
 		],
 	);
